// src/app/model/proposition.model.ts
export class Proposition {

    idDispo?: any;
	idUserInitiateur?: string;	
	idVol?: string;
	villeDepart?: string;
	aeronomDepart?: string;
	dateDepart?: Date;
	villeArrivee?: string;
	aeronomArrivee?: string;
	dateArrivee?: Date;
	idUserCandidat?: string;
	nbKiloAchete?: any;
	idStatut?: any;
	libelStatut?: string;

}